feat: Eventual Consistency + 비보상형 Saga 구현
트렌드 키워드는 무조건 저장하고,
채팅방 생성 실패 시 rollback하지 않고,
재시도하거나 DLQ에 쌓았다가 추후 복구

이 구조는 정합성은 결국 맞춰짐(eventually consistent)

하지만 중간에 실패가 발생해도 고가치 데이터 유실이 없음
