--- trend-service/src/main/java/com/trendchat/trendservice/service/TrendKeywordProducer.java
@@ -1,17 +1,0 @@
-package com.trendchat.trendservice.service;
-
-import com.trendchat.trendservice.dto.TrendKeywordItem;
-import lombok.RequiredArgsConstructor;
-import org.springframework.kafka.core.KafkaTemplate;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class TrendKeywordProducer {
-
-    private final KafkaTemplate<String, Object> kafkaTemplate;
-
-    public void sendKeyword(String key, TrendKeywordItem value) {
-        kafkaTemplate.send("trend-keywords", key, value);
-    }
-}
